gerando tabelas automaticamente com jpa

+++ src/main/java/com/guilhermeferreira/livrariaapi/model/Livro.java
@@ -35,7 +35,7 @@ public class Livro {
     private GeneroLivro genero;
 
     @Column(name = "preco",precision = 18, scale = 2)
-    private Double preco;
+    private BigDecimal preco;
    // private BigDecimal preco;
 
     @ManyToOne
